requirements.txt
phidata 
openai 
duckduckgo-search
packaging
yfinance
groq
exa_py
streamlit